requirements.txt
Flask>=3.0
flask-cors>=4.0
gunicorn>=21.2.0
google-auth>=2.35
google-auth-oauthlib>=1.2
google-auth-httplib2>=0.2.0
google-api-python-client>=2.149
openai>=1.52
python-dotenv>=1.0
